01 - Reading List: Moved layouts folder into components folder.
