+++ pruebas/01-reading-list/iivvaannxx/src/features/library/components/layouts/grid/icon.svelte
